Bit o re-arranging
