+++ packages/text-to-code/src/text_to_code/services/result_cache.py
@@ -1,69 +1,10 @@
 import json
-from datetime import UTC
-from datetime import datetime
 from hashlib import sha256
 
 from opensearchpy import OpenSearch
-from pydantic import Field
 
 from shared_models import Code
-from shared_models import FrozenBaseModel
-
-
-class OpenSearchResultCacheSource(FrozenBaseModel):
-    """Represents a term-standardization mapping returned from the OpenSearch Result Cache Index."""
-
-    cache_key: str = Field(
-        description="The computed hash key mapping candidate text + data field to this known hit."
-    )
-    text: str = Field(description="The input candidate text this Cache Result is associated with.")
-    data_field: str = Field(
-        description="The data field from which the candidate text was retrieved during initial "
-        "processing of this cache hit."
-    )
-    loinc_code: str = Field(description="The serialized Code object for this cached input.")
-    search_score: float = Field(
-        description="The cosine similarity score calculated for this standardization during "
-        "initial processing of this input."
-    )
-    reranker_score: float = Field(
-        description="The cross-encoder score calculated for this standardization during initial "
-        "processing of this input."
-    )
-    cached_at: str = Field(description="The timestamp at which this input was cached.")
-
-
-class OpenSearchGetResponse(FrozenBaseModel):
-    """Represents the response body returned by an ID-indexed GET request to the OpenSearch API."""
-
-    index: str = Field(description="The name of the index containing the document.", alias="_index")
-    id: str = Field(description="The document's unique identifier.", alias="_id")
-    version: int = Field(
-        description="The document's version number, incremented each time the document is updated",
-        alias="_version",
-    )
-    seq_no: int = Field(
-        description="The sequence number assigned to the document for the indexing operation. "
-        "Used to ensure an older version doesn't overwrite a newer version.",
-        alias="_seq_no",
-    )
-    primary_term: int = Field(
-        description="The primary term assigned to the document for the indexing operation. "
-        "Used with `seq_no` for optimistic concurrency control.",
-        alias="_primary_term",
-    )
-    found: bool = Field(
-        description="Indicates whether the document exists (`True` if it was found)."
-    )
-    routing: str = Field(
-        description="The routing value used to determine which shard stores the document. "
-        "Only included if a routing value was specified when the document was indexed.",
-        alias="_routing",
-    )
-    source: OpenSearchResultCacheSource = Field(description="The cached result for the input.")
-    fields: dict[str, list] = Field(
-        description="Stored field values of the mappings.", alias="_fields"
-    )
+from text_to_code.models.result_cache import OpenSearchResultCacheSource
 
 
 def get_cached_result(
@@ -123,7 +64,6 @@ def put_new_cached_result(
         loinc_code=json.dumps(loinc_code.__dict__),
         search_score=search_score,
         reranker_score=reranker_score,
-        cached_at=datetime.now(UTC).isoformat(),
     )
     put_response = opensearch_client.index(index=index, id=cache_key, body=new_cache_hit)
     return put_response["result"] == "created"
